third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/styles-disable-property-after-selector-edit.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function test() {
  TestRunner.runTestSuite([
    function selectInitialNode(next) {
      ElementsTestRunner.selectNodeAndWaitForStyles('inspected', next);
    },

    function editSelector(next) {
      var section = ElementsTestRunner.firstMatchedStyleSection();
      section.startEditingSelector();
      section._selectorElement.textContent = '#inspected, .INSERTED-OTHER-SELECTOR';
      ElementsTestRunner.waitForSelectorCommitted(next);
      section._selectorElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
    },

    function testDisableProperty(next) {
      var treeItem = ElementsTestRunner.getMatchedStylePropertyTreeItem('color');
      ElementsTestRunner.waitForStyleApplied(onPropertyDisabled);
      treeItem._toggleEnabled({target: {checked: false}, consume: function() {}});

      function onPropertyDisabled() {
        TestRunner.addResult('\n\n#### AFTER PROPERTY DISABLED ####\n\n');
        ElementsTestRunner.dumpSelectedElementStyles(true, false, true);
        next();
      }
    }
  ]);
}
</script>
<style>
#inspected {
    color: red;
}
</style>
</head>

<body onload="runTest()">
<p>
Verifies that sequence of setting selector and disabling property works.
</p>

<div id="inspected">Red text here.</div>

</body>
</html>
