third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/styles-disable-property-after-selector-edit-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verifies that sequence of setting selector and disabling property works.


Running: selectInitialNode

Running: editSelector

Running: testDisableProperty


#### AFTER PROPERTY DISABLED ####


[expanded] 
element.style { ()

[expanded] 
#inspected, .INSERTED-OTHER-SELECTOR { (<style>…</style>)
/-- overloaded --/ /-- disabled --/     /* color: red; */

[expanded] 
div { (user agent stylesheet)
    display: block;
